+++ Changelog.md
@@ -2,6 +2,10 @@
 
 ## Current Master
 
+- Nothing yet!
+
+## 0.12.0
+
 - Utilize `included` rule when finding files. See [#79](https://github.com/ashfurrow/danger-ruby-swiftlint/pull/79).
 
 ## 0.11.1